Refactor code
Cleaning up code, moving to more relevant packages and better names. Fixing late night choices. Adding getters to Views
