+++ src/main/java/com/chess/api/core/exception/AppErrorConfiguration.java
@@ -1,4 +1,4 @@
-package com.chess.api;
+package com.chess.api.core.exception;
 
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.web.servlet.error.ErrorAttributes;
@@ -6,14 +6,14 @@ import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 
 @Configuration
-public class ApiErrorConfiguration {
+public class AppErrorConfiguration {
 
     @Value("${api.version}")
     private String currentApiVersion;
 
     @Bean
     public ErrorAttributes errorAttributes() {
-        return new ApiErrorAttributes(this.currentApiVersion);
+        return new AppErrorAttributes(this.currentApiVersion);
     }
 
 }
