+++ src/main/java/com/chess/api/view/response/LoginView.java
@@ -1,8 +1,10 @@
 package com.chess.api.view.response;
 
 import lombok.AllArgsConstructor;
+import lombok.Getter;
 
 @AllArgsConstructor
+@Getter
 public class LoginView {
 
     private final String jwt;
