+++ src/main/java/com/chess/api/view/response/PieceView.java
@@ -3,8 +3,10 @@ package com.chess.api.view.response;
 import com.chess.api.data.piece.Point;
 import java.util.Set;
 import lombok.AllArgsConstructor;
+import lombok.Getter;
 
 @AllArgsConstructor
+@Getter
 public class PieceView {
 
     private final String code;
