+++ src/main/java/com/chess/api/view/response/PlayerView.java
@@ -1,8 +1,10 @@
 package com.chess.api.view.response;
 
 import lombok.AllArgsConstructor;
+import lombok.Getter;
 
 @AllArgsConstructor
+@Getter
 public class PlayerView {
 
     private final String username;
